+++ src/app/controllers/postsCtrl.js
@@ -1,0 +1,9 @@
+(function() {
+  'use strict';
+  var wsNinja = angular.module('app');
+
+  wsNinja.controller('PostsListCtrl', function($scope, Posts) {
+    $scope.posts = Posts.list();
+  });
+
+})();