+++ src/app/controllers/projectsCtrl.js
@@ -1,0 +1,8 @@
+(function() {
+  'use strict';
+  var wsNinja = angular.module('app');
+
+  wsNinja.controller('ProjectsListCtrl', function($scope, Projects) {
+    $scope.projects = Projects.list();
+  });
+})();